Add react-mobile-cropper
This library is based on `react-advanced-cropper`. Unlike the original library this library doesn't focus on flexibility, it's ready to use solution with specific behavior and interface.

[Documentation](https://github.com/advanced-cropper/react-mobile-cropper) / [sandbox](https://codesandbox.io/s/react-mobile-cropper-c1bz2)

According to the rules, I've removed the link to [`react-json-tree`](https://github.com/alexkuz/react-json-tree), because it's archived now and the repository content moved to `redux-devtools` monorepository that is already in this list.